@@ -3,7 +3,7 @@
 namespace Port\Tests\Filter;
 
 use PHPUnit\Framework\TestCase;
-use PHPUnit_Framework_MockObject_MockObject as Mock;
+use PHPUnit\Framework\MockObject\MockObject as Mock;
 use Symfony\Component\Validator\ConstraintViolationInterface;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\Constraints;
@@ -19,7 +19,7 @@ class ValidatorFilterTest extends TestCase
     /** @var Mock|ValidatorInterface */
     private $validator;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->validator = $this->createMock(ValidatorInterface::class);
         $this->filter = new ValidatorFilter($this->validator);
