@@ -12,6 +12,6 @@ class DuplicateHeadersException extends ReaderException
      */
     public function __construct(array $duplicates)
     {
-        parent::__construct(sprintf('File contains duplicate headers: %s', implode($duplicates, ', ')));
+        parent::__construct(sprintf('File contains duplicate headers: %s', implode(', ', $duplicates)));
     }
 }
