@@ -2,11 +2,13 @@
 
 namespace Port;
 
+use Iterator;
+
 /**
  * Iterator that reads data to be imported
  *
  * @author David de Boer <david@ddeboer.nl>
  */
-interface Reader extends \Iterator
+interface Reader extends Iterator
 {
 }
