@@ -2,11 +2,13 @@
 
 namespace Port\Reader;
 
+use ArrayIterator;
+
 /**
  * Reads an array
  *
  * @author David de Boer <david@ddeboer.nl>
  */
-class ArrayReader extends \ArrayIterator implements CountableReader
+class ArrayReader extends ArrayIterator implements CountableReader
 {
 }
