@@ -3,6 +3,7 @@
 namespace Port\ValueConverter;
 
 use Port\Exception\UnexpectedTypeException;
+use RuntimeException;
 
 /**
  * Convert a value in a specific charset
@@ -44,6 +45,6 @@ class CharsetValueConverter
             return iconv($this->inCharset, $this->charset, $input);
         }
 
-        throw new \RuntimeException('Could not convert the charset. Please install the mbstring or iconv extension!');
+        throw new RuntimeException('Could not convert the charset. Please install the mbstring or iconv extension!');
     }
 }
