@@ -1,13 +1,16 @@
 <?php
 
-namespace Port\Filter;
+namespace Port\Tests\Filter;
 
+use DateTime;
+use LogicException;
 use PHPUnit\Framework\TestCase;
+use Port\Filter\DateTimeThresholdFilter;
 use Port\ValueConverter\DateTimeValueConverter;
 
 class DateTimeFilterTest extends TestCase
 {
-    public function setUp()
+    public function setUp(): void
     {
         $this->items = array(
             'a' => array('updated_at' => '-3 day'),
@@ -24,12 +27,10 @@ class DateTimeFilterTest extends TestCase
         return array_filter($items, array($filter, '__invoke'));
     }
 
-    /**
-     * @expectedException \LogicException
-     * @expectedExceptionMessage Make sure you set a threshold
-     */
     public function testDefaultFilter()
     {
+        $this->expectExceptionMessage("Make sure you set a threshold");
+        $this->expectException(LogicException::class);
         $this->applyFilter(
             new DateTimeThresholdFilter(new DateTimeValueConverter()),
             $this->items
@@ -40,7 +41,7 @@ class DateTimeFilterTest extends TestCase
     {
         $resultItems = $this->applyFilter(new DateTimeThresholdFilter(
             new DateTimeValueConverter(),
-            new \DateTime('today')
+            new DateTime('today')
         ), $this->items);
         $this->assertEquals(
             array('d', 'e', 'f'),
@@ -51,7 +52,7 @@ class DateTimeFilterTest extends TestCase
     public function testSetter()
     {
         $filter = new DateTimeThresholdFilter(new DateTimeValueConverter());
-        $filter->setThreshold(new \DateTime('today'));
+        $filter->setThreshold(new DateTime('today'));
         $resultItems = $this->applyFilter($filter, $this->items);
 
 
