@@ -2,8 +2,11 @@
 
 namespace Port\Tests;
 
+use DateTime;
+use Exception;
 use PHPUnit\Framework\TestCase;
 use Port\Result;
+use SplObjectStorage;
 
 /**
  * Tests For Workflow Result
@@ -14,21 +17,21 @@ class ResultTest extends TestCase
 {
     public function testResultName()
     {
-        $result = new Result('export', new \DateTime, new \DateTime, 10, new \SplObjectStorage());
+        $result = new Result('export', new DateTime, new DateTime, 10, new SplObjectStorage());
         $this->assertSame('export', $result->getName());
     }
 
     public function testResultCounts()
     {
-        $result = new Result('export', new \DateTime, new \DateTime, 10, new \SplObjectStorage());
+        $result = new Result('export', new DateTime, new DateTime, 10, new SplObjectStorage());
         $this->assertSame(10, $result->getTotalProcessedCount());
         $this->assertSame(10, $result->getSuccessCount());
         $this->assertSame(0, $result->getErrorCount());
 
-        $exceptions = new \SplObjectStorage();
-        $exceptions->attach(new \Exception());
-        $exceptions->attach(new \Exception());
-        $result = new Result('export', new \DateTime, new \DateTime, 10, $exceptions);
+        $exceptions = new SplObjectStorage();
+        $exceptions->attach(new Exception());
+        $exceptions->attach(new Exception());
+        $result = new Result('export', new DateTime, new DateTime, 10, $exceptions);
         $this->assertSame(10, $result->getTotalProcessedCount());
         $this->assertSame(8, $result->getSuccessCount());
         $this->assertSame(2, $result->getErrorCount());
@@ -37,10 +40,10 @@ class ResultTest extends TestCase
 
     public function testDates()
     {
-        $startDate  = new \DateTime("22-07-2014 22:00");
-        $endDate    = new \DateTime("22-07-2014 23:30");
+        $startDate  = new DateTime("22-07-2014 22:00");
+        $endDate    = new DateTime("22-07-2014 23:30");
 
-        $result     = new Result('export', $startDate, $endDate, 10, new \SplObjectStorage());
+        $result     = new Result('export', $startDate, $endDate, 10, new SplObjectStorage());
 
         $this->assertSame($startDate, $result->getStartTime());
         $this->assertSame($endDate, $result->getEndTime());
@@ -49,27 +52,27 @@ class ResultTest extends TestCase
 
     public function testHasErrorsReturnsTrueIfAnyExceptions()
     {
-        $exceptions = new \SplObjectStorage();
-        $exceptions->attach(new \Exception());
-        $exceptions->attach(new \Exception());
+        $exceptions = new SplObjectStorage();
+        $exceptions->attach(new Exception());
+        $exceptions->attach(new Exception());
 
-        $result = new Result('export', new \DateTime, new \DateTime, 10, $exceptions);
+        $result = new Result('export', new DateTime, new DateTime, 10, $exceptions);
         $this->assertTrue($result->hasErrors());
     }
 
     public function testHasErrorsReturnsFalseIfNoExceptions()
     {
-        $result = new Result('export', new \DateTime, new \DateTime, 10, new \SplObjectStorage());
+        $result = new Result('export', new DateTime, new DateTime, 10, new SplObjectStorage());
         $this->assertFalse($result->hasErrors());
     }
 
     public function testGetExceptions()
     {
-        $exceptions = new \SplObjectStorage();
-        $exceptions->attach(new \Exception());
-        $exceptions->attach(new \Exception());
+        $exceptions = new SplObjectStorage();
+        $exceptions->attach(new Exception());
+        $exceptions->attach(new Exception());
 
-        $result = new Result('export', new \DateTime, new \DateTime, 10, $exceptions);
+        $result = new Result('export', new DateTime, new DateTime, 10, $exceptions);
         $this->assertSame($exceptions, $result->getExceptions());
     }
 }
