@@ -4,6 +4,7 @@ namespace Port\Tests\ValueConverter;
 
 use PHPUnit\Framework\TestCase;
 use Port\ValueConverter\DateTimeValueConverter;
+use UnexpectedValueException;
 
 class DateTimeValueConverterTest extends TestCase
 {
@@ -46,7 +47,8 @@ class DateTimeValueConverterTest extends TestCase
     {
         $value = '14/10/2008 09:40:20';
         $converter = new DateTimeValueConverter('d-m-y', 'd-M-Y');
-        $this->expectException("UnexpectedValueException", "14/10/2008 09:40:20 is not a valid date/time according to format d-m-y");
+        $this->expectException(UnexpectedValueException::class);
+        $this->expectExceptionMessage("14/10/2008 09:40:20 is not a valid date/time according to format d-m-y");
         call_user_func($converter, $value);
     }
 
