@@ -2,9 +2,12 @@
 
 namespace Port\Tests\Writer;
 
+use InvalidArgumentException;
+use stdClass;
+
 class AbstractStreamWriterTest extends StreamWriterTest
 {
-    protected function setUp()
+    protected function setUp(): void
     {
         $this->writer = $this->getMockForAbstractClass('Port\\Writer\\AbstractStreamWriter');
     }
@@ -16,13 +19,13 @@ class AbstractStreamWriterTest extends StreamWriterTest
 
     public function testItThrowsInvalidArgumentExceptionOnInvalidStream()
     {
-        $invalidStreams = array(0, 1, null, 'stream', new \stdClass());
+        $invalidStreams = array(0, 1, null, 'stream', new stdClass());
         foreach ($invalidStreams as $invalidStream) {
             try {
                 $this->writer->setStream($invalidStream);
                 $this->fail('Above call should throw exception');
-            } catch (\InvalidArgumentException $exception) {
-                $this->assertContains('Expects argument to be a stream resource', $exception->getMessage());
+            } catch (InvalidArgumentException $exception) {
+                $this->assertStringContainsString('Expects argument to be a stream resource', $exception->getMessage());
             }
         }
     }
